refactor(types): 14 consumer 移行 + StepKind alias + StepNote 移管 (#1186 Phase 2-B)
ISSUE #1186 Phase 2-B: action.ts 経由 import の段階移行。安全な leaf consumer 14 file
を types/v3 + utils/processFlowMetadata 直接 import に置き換え。

## 構造変更

- types/v3/process-flow.ts: `StepKind = Step["kind"]` named export 追加 (consumer 利便性)
- utils/processFlowMetadata.ts: StepNote / StepNoteType / STEP_NOTE_TYPE_VALUES を action.ts から移管
  - 理由: UI 表示 + migration 互換が frontend 固有概念 (common.v3 Note には legacy `type?` field なし)
- types/action.ts: 上記移管項目を re-export only に縮退 (backward compat 維持)

## 移行した consumer (14 file)

| file | 移行内容 |
|---|---|
| ProcessFlowEditor.tsx | ProcessFlow/Step/ActionTrigger → v3、STEP_TEMPLATES/STEP_TYPE_LABELS → metadata |
| ExternalOutcomesPanel.tsx | ExternalSystemStep/Step → v3、ExternalCallOutcome/VALUES → metadata |
| MaturityBadge.tsx | Maturity → v3 |
| NotesPanel.tsx | StepNote/StepNoteType → metadata、STEP_NOTE_TYPE_VALUES → metadata |
| PalettePanel.tsx | StepType → v3 (StepKind alias)、STEP_TEMPLATES → metadata |
| StepContextMenu.tsx | StepType → v3、STEP_TYPE_* → metadata |
| PaletteButtons.tsx | StepType → v3、STEP_TYPE_* → metadata |
| stepCardConstants.ts | StepType → v3、DbOperation は local type 化 |
| AddActionModal.tsx | ActionTrigger → v3 |
| WorkflowStepPanel.tsx | Step/WorkflowStep/WorkflowPattern → v3、WORKFLOW_PATTERN_* → metadata |
| StepCard.tsx | ProcessFlow/Step/StepType → v3、STEP_TYPE_*/WORKFLOW_PATTERN_LABELS → metadata |
| SortableStepCard.tsx | ProcessFlow/Step/StepType → v3 (StepCard と整合) |
| TransactionScopeStepPanel.tsx | Step/StepType/TransactionScopeStep/ProcessFlow → v3、STEP_TYPE_* → metadata |
| specExporter.test.ts | 全 step variant → v3 (OtherStep のみ action.ts 残置) |

## 残置 (Phase 2-C/D 以降で対応)

WorkflowApprover / WorkflowQuorum / TransactionIsolationLevel / TransactionPropagation /
ExternalCallOutcomeSpec / OtherStep 等は action.ts に AnyRecord として残置 (v3 strict 化は次フェーズ)。

action.ts の旧 import path を引き続き使う consumer は backward compat re-export 経由で
変更なく動作 (残り 85 file)。

## Test plan

- [x] frontend build (tsc + vite) → OK
- [x] frontend test → 2346 / 2346 pass
- [x] backend test → 516 / 516 pass
- [x] validate:samples (retail / realestate) → 7+1 flows pass

Refs #1186
